Move: GeminiClient 폴더 생성 및 이동
