+++ src/services/gemini-client/tests/gemini-client.test.ts
