+++ src/services/gemini-client/tests/integration.test.ts
